+++ RELEASENOTES.md
@@ -99,6 +99,8 @@
 *   Remove deprecated symbols:
     *   Remove `androidx.media3.extractor.metadata.flac.VorbisComment`. Use
         `androidx.media3.extractor.metadata.vorbis.VorbisComment` instead.
+    *   Remove `androidx.media3.extractor.metadata.mp4.MotionPhotoMetadata`. Use
+        `androidx.media3.extractor.metadata.MotionPhotoMetadata` instead.
     *   Remove `MediaSource.prepareSource(MediaSourceCaller, TransferListener)`.
         Use `MediaSource.prepareSource(MediaSourceCaller, TransferListener,
         PlayerId)` instead.